Move e2e_test.go to e2e.go
e2e.go was just a stub file and when you have a filename appended
with _test.go, that file will not be visible at compile time which
gets in the way of using functions, constants, etc in other files.
one example was util.go trying to use podNetworkAnnotation.
